fix: load node_modules correctly
When I added exposed the node_modules attribute on nodejs_binary in
bcf19e0200fe44a2d03443b5303fb724fb112f30
I didn't test this with the case where a rule in a subdirectory uses the implicit @//:node_module
in a parent directory. We should use the location of the :node_modules label, not the nodejs_binary
rule.