Adjust to a breaking change in rules_nodejs.
https://github.com/bazelbuild/rules_nodejs/commit/96ecb35549098a341c9db3ce7c578ef4788b67f1#diff-c3f5cd9c76ccb895b38b4cb04332e2aa
introduced a subtle breaking change. Instead of loading dependencies from the runfiles associated with the workspace declaring the rule, we now get them from the workspace calling the rule. That's better in general, but it means that nodejs_binary users need to be explicit about which node_modules to use.

PiperOrigin-RevId: 174403902