More bugfixes for naming AMD modules following the module_name of a ts_library.
The naming scheme now requires local knowledge: within the compilation of a given
library, we know what module name it wants. But when we load a .d.ts file from disk,
we don't know what module name was used to compile that file.

It's possible we could use the tsconfig paths setting to re-derive this knowledge,
but then we must understand pathMapping semantics, such as dealing with trailing /*.

Instead, relax our logic so we only attempt to give names to non-.d.ts source files.
The prior logic was more strict only out of caution, not because it was required.

Also, as a precaution don't try to name modules if a file is not underneath the bazel
package we are currently building (this is a redundant check; the module mapping logic
should disallow including a file outside the package).

PiperOrigin-RevId: 187635489