bump typescript worker cache size
In practice, we cannot really use the memory usage to guide whether
to cache or not -- the memory usage only goes down after a gc (which is
costly to run) and it's also at the mercy of whether v8 actually wants to
return its unused memory to the OS and so on.  Observationally the memory
usage is pretty unpredictable and we should not use it at all.

But there's one especially pathological case, which is where a single
compilation causes the worker process to go past its memory limit.  In
that case we start trying to bail out files that the current compilation
is using from the cache, which causes us to completely empty the cache.
This problem occurs more often with iclutz, which feeds a lot more d.ts
to the compiler.

To avoid this problem without a more invasive change, just bump the cache
limit up to a larger value.  The nodejs process itself with just tsc loaded
(before we run any compiles) is around 250mb, so a 300mb limit is way too
low.

PiperOrigin-RevId: 188836865