Fix crash of tsc_wrapped.
If tsickle reports a warning, we try to read the ignoreWarningPaths property of the BazelOptions. The type of this property is non-optional, so we should always set a value when creating a bazelOpts dict.

Ideally we should have a JSON schema for the tsconfig.json and run some validator to assert we have an object that satisfies our interface.

PiperOrigin-RevId: 196295708