Enable noErrorTruncation for TypeScript builds in google3.
For some errors, the truncation can make the error condition impossible to understand, e.g. when abbreviating the set of union types in some expression. While very long errors might get harder to read, we expect those not to occur too often.

It's hard to predict this, so we'll enable this for a while and see what the effect is.

PiperOrigin-RevId: 199681601