Don't add workspace to expanded ../ paths
Otherwise they will become 'workspace/../path', which is the same as just 'path'.

Although both are valid, these paths are meant to be looked up in the runfiles so they should be normalized already instead of leavinf it for the consuming tool.
This is especially important in windows where the runfiles manifest contains strings of normalized paths.