fix: local workspace path not always properly determined
Since there can be also runfiles which refer to the current `USER_WORKSPACE`, but actually resolve to the bazel output instead of to the actual local workspace, it can happen that the `node_loader` determines a wrong path to the local workspace.

We can safely fix this by just ensuring that the runfile, which potentially resolves to the local workspace, does not refer to the bazel-bin or bazel-genfiles directory (those paths are templated)

Fixes #352.