Disallow TypeScript-producing rules as srcs in ts_library and ng_module.
Previously, it was possible to write a TypeScript rule depending on
another TypeScript rule:

    ts_library(name = "a", srcs = ["a.ts"])
    ts_library(name = "b", srcs = [":a"])

Because the default "files" output of "a" above is "a.d.ts", "b" will
accept this input and treat it as a regular ".d.ts" source file.

That's very confusing, but also wrong on multiple levels: it'll treat
the symbols from the .d.ts file as external, preventing renaming and
generating externs. It also has a change to subtly break module
resolution, and generally confuse tsickle.

PiperOrigin-RevId: 221794400