Do not create empty BUILD files in packages that do not contain any TS.
In a previous change, we changed taze not to track whether it made
changes to the BUILD file, but rather just make the changes and compare
the BUILD file(s) for equality after the fact. That change had a bug
where it'd always create a (potentially empty) BUILD file, even if non
previously existed and the package does not contain any TypeScript.

This change fixes the problem by treating a newly created empty BUILD
file as equivalent to no BUILD file at all, and not writing the file in
that case.

PiperOrigin-RevId: 222072180