Increase `npm/yarn install` timeout and add new arg for `--network-timeout`
Previously, the `timeout` arg was used for both the
`repository_ctx.execute()` timeout and the yarn's `--network-timeout`.
Yet, in some cases the default 10mins timeout is not enough for the
`npm/yarn install` job to complete (e.g. on CI VM with slow network
connection). Additionally, it makes little sense to set the per-request
network timeout and the overall command timeout (which potentially
involves multiple network requests) to the same value.

This commit introduces a new arg for `yarn_install()` to set
`--network-timeout` (defaulting to 5mins) and increases the default
`timeout` value to 60mins (from 10mins).