Improve handling of generated source files.
Previously, the query-based analyzer looked for the dep providing a generated source file by walking up the graph of generating rules, which relied on generated files only happening in macros and the ts_library in the macro sharing a label with the macro.  Instead, generated sources should be handled like any other source - the analyzer should look for the generated file (or the generating rule) in the srcs attribute of some other rule.

PiperOrigin-RevId: 231693464