fix(devserver): static assets sometimes incorrectly injected
Currently whenever the `ts_devserver` tries to inject static
assets which are not part of the current Bazel package, but
part of the current workspace (e.g. file in a parent package),
incorrect asset URLs are injected.

This is because the `ts_devserver` currently uses the `File#path`
property which is relative to the execroot. So files which are not
part of the current Bazel package, but part of the execroot sources,
will be kept relative to the execroot. This is problematic because
the execroot is not part of the devserver root directories and
therfore the injected URLs are incorrect. Instead we should always
explicitly pass the `ts_devserver` implementation-specific manifest
paths to the asset injection function.

To be clear: This currently only affects Windows because if Bazel
symlinks the runfiles, the Go runfile helpers also look for the
requested files in the `working dirctory` (which is always the
the `devserver_runfiles/{workspace_name}` --> so runfiles within
the current workspace can be always resolved even though they
are not part of the root directory. This is something we need
to make consistent in the future, but is not related to that
specific bug (we should always precisely pass manifest paths)

Fixes #409.

Closes #423

PiperOrigin-RevId: 234678783