Rename NodeModuleSources to NodeModulesSourcesInfo per naming convention
BREAKING CHANGE:

Although this name is an internal detail and is not used anywhere outside of this repository, since we publish multiple packages this change will still mean that an older @bazel/typescript npm package will not work with the next release of build_bazel_rules_nodejs and vice-versa.

Next release should be a minor bump with this.