Rollup bundle rule does not resolve external source maps
In order to support external source-maps, the `rollup-plugin-sourcemaps` module has been added to the
rollup configuration. Unfortunately this currently only has an effect on Windows as the sandbox is not
available and the plugin can read the referenced external `.map` files.

When running in the Bazel sandbox or within RBE, these map files can't be read by the plugin because the `.map`
files are omitted from the `run_rollup` action inputs. This causes inconsistent behavior and makes the sourcemaps
plugin a no-op right now.