feat(jasmine_node_test): add attr `jasmine_config`
This commit enables the `jasmine_node_test` rule to be configured on a
per-test basis through jasmine's configuration file.

The demanding situation for me was to turn off randomnization (default
to on in Bazel) to guarantee some test results.  This would incur adding
the following attribute to jasmine's configuration file:
    ```json
    {
        "random": false
    }
    ```