Fix for --incompatible_depset_is_not_iterable
We used a depset for a collection which is only used within the context of a single target: the files[] block in the generated tsconfig. As documented on https://docs.bazel.build/versions/master/skylark/depsets.html:
"If you don’t need the merge operation, consider using another type, such as list or dict."

The problem is exposed by --incompatible_depset_is_not_iterable which fails at the spot where we iterated the depset of files, showing that it should have been a list.

Also clean up a warning about the deprecated load of jasmine_node_test

See #443

Closes #451

PiperOrigin-RevId: 249758830