Remove the tsconfig override warning feature.
It is mostly unavoidable: users need pathMapping to make their editor happy with multiple named libs in a monorepo, some language features require specifying target, etc.

Also it's reporting to every user on every compile, this is not usually the right time to fix it.

If we want to rollforward a feature, it should be based more precisely on things that confused users, and should also take care to only warn if the users setting conflicts with Bazel's in a breaking way.

Alternative to https://github.com/angular/angular/pull/30104

PiperOrigin-RevId: 250324968