fix(typescript): add peerDependency on typescript
Set it to 3.0 or greater. This lets us drop some e2e testing which was taking a long time on CI.
Also add e2e testing for TS 3.6 which was just released.

Note that typescript doesn't do semver so it's possible that something breaks before 4.0. If users report that, we'll have to re-publish with a narrowed range.