Default rollup bundle config specifies invalid input option
Apparently the default rollup bundle configuration specifies an invalid option called `resolveBazel`. This causes an error to be thrown as rollup complains about invalid/unknown options.

This is not an issue for the default consumers of the rollup bundle configuration as those specify `--silent`, but it will be an issue if someone uses its own rollup rule that uses this configuration. Specifically I hit this scenario on windows where the arguments for the rollup rule were too long and the "--silent" option got accidentally omitted..