chore: deprecate old rollup_bundle rule
At this point we have rolled out the new one everywhere downstream so we are confident it works for everyone.
Also delete our tests for the old rule, rather than amend our CI to enable the legacy rule.
We don't plan to make any more changes so we don't need the tests.