fix(builtin): allow more than 2 segments in linker module names
The linking code is made more robust to handle more than two segments in the module names. There is no technical reason to not allow this although it does diverge from the npm package standard of either foo or @scope/foo names.

Module names are sorted and grouped to ensure that nested modules such as ‘c’ are symlinked first before ‘c/c’, ‘c/c/c/c’, etc and that groups of nested modules are symlinked in order while maintaining the parallelized symlink code for module names that are not nested. Gaps in the nested paths are filled with mkdir as needed and common paths between groups are mkdir'd before handing off to the parallelized symlink code.