refactor: change generated bin loadpoints to shorter name
Users are confused by the index.bzl filename: where does it come from?
For vanity, the load point of @npm//mocha/bin is so short, it looks good

Note this assumes no package published a bin named bin - in that case Bazel will error that the new alias defines a duplicate target. I suppose we'll have to special case that if it happens.