feat(karma): npm peer deps & remove @rules_webtesting//browsers/sauce:chrome-win10 support
Also cleanup us arguments and doc string for karma_web_test_suite.

BREAKING CHANGES:

npm karma deps for karma_web_test and karma_web_suite are now peer deps so that the versions used can be chosen by the user.

This PR also removes the built-in  `@io_bazel_rules_webtesting//browsers/sauce:chrome-win10` saucelabs support. It is not very useful as it only tests a single browser and it difficult to use. In the angular repo, saucelabs support was implemented with a custom karma config using karma_web_test. This is the recommended approach.