fix(bazel): no longer always include the version_file in inputs
While bazel properly handles not causing rebuilds based on version_file locally,
if the version_file is included in the inputs it prevents caching from ever
working when caching using remote cache or disk cache.

Instead of determining if stamping should occur based on the presence of the
version_file, we should choose to stamp based on an attribute stamp.