refactor(builtin): renamed npm_package to pkg_npm to match naming convention
It also frees up the name `npm_package` which may be useful in the future as a non-packaging representation of an npm package as an alternative to promoting js_library to the public API.

BREAKING CHANGE:
`npm_package` renamed to `pkg_npm`. This is to match the naming convention for package rules https://docs.bazel.build/versions/master/be/pkg.html.