fix(builtin): nodejs_binary collects module_mappings for linker
Module name cannot be equal to the workspace name with the linker so this change removes the regression test for 834. This was a corner case and was already broken with the 1.0 for npm_package_bin so this fix makes nodejs_binary consistent with npm_package_bin and the new desired behavior for 1.0.