fix(examples): fix karma config in examples/angular_bazel_architect
Use ChromeHeadess as base instead of passing —headless. This fixes the CI failure on CircleCI and fixes it locally for me on OSX. buildkite macos still fails with

```
        # ==================== Test output for //:test:
        # 29 01 2020 23:31:42.804:INFO [karma-server]: Karma v4.1.0 server started at http://0.0.0.0:9876/
        # 29 01 2020 23:31:42.806:INFO [launcher]: Launching browsers ChromeHeadlessNoSandbox with concurrency unlimited
        # 29 01 2020 23:31:42.808:INFO [launcher]: Starting browser ChromeHeadless
        # 29 01 2020 23:32:42.815:WARN [launcher]: ChromeHeadless have not captured in 60000 ms, killing.
        # 29 01 2020 23:32:44.818:WARN [launcher]: ChromeHeadless was not killed in 2000 ms, sending SIGKILL.
        # 29 01 2020 23:32:46.822:WARN [launcher]: ChromeHeadless was not killed by SIGKILL in 2000 ms, continuing.
```

but this seems to be something to do with the buildkite machine chrome version/configuration as it does work locally on mac for me with Chrome 79. Disabled the test on using `no-bazelci-mac` tag.