feat(builtin): expose `@npm//foo__all_files` filegroup that includes all files in the npm package
Including files with spaces & unicode characters & files that have been filtered out by the `included_files` attribute. This filegroup cannot be used in runfiles because of Bazel issue https://github.com/bazelbuild/bazel/issues/4327, but it can be used for other purposes such as the srcs input to a pkg_tar for generating a tar of an npm package pulled down with yarn_install or npm_install.