chore: warn on usage of install_bazel_dependencies
This function is no longer needed since #1149 was finished. However, @angular/bazel relies on it, as might other third-party packages we know nothing about.
The conservative rollout is to warn on usage for now, then remove the feature in 3.0

This change also removes our own usage of the function except for angular_view_engine. We also stop writing it in new workspaces created with @bazel/create