docs: escape HTML entities
Stardoc extracts the documentation from .bzl files and generate them
as Markdown, but some contents are in HTML table. For such content,
we need to escape HTML entities otherwise they'll be rendered as
HTML node. This not only messes up the actual docs page, but could also
lead to potential security vulnerability.
For example, see the `format` attribute of `rollup_bundle` on
https://bazelbuild.github.io/rules_nodejs/Rollup.html (end of page).