fix: support running with --enable_runfiles on Windows
I was finding the run and test commands on a nodejs binary on
Windows gave an error message similar to the one mentioned in the
following post when --enable_runfiles is passed in on the command line:

https://github.com/bazelbuild/rules_nodejs/pull/2178#issuecomment-694146500

Running with --noenable_runfiles, the tests worked correctly, but
other packages in my repo require runfiles.

This patch changes the resolution to use the MANIFEST file inside
the runfiles folder when runfiles are enabled, which seems to fix
the issue for me.