refactor: move protobufjs to userspace protobufjs_library macro
The Labs package has this old and crufty protobufjs_ts_library rule which is not really documented.
The grpc-based one was meant to replace it.

Instead of providing a rule in @bazel/labs, just add an example showing how this can be done in userspace with a macro.
(there's also a one-liner rule to convert ProtoInfo to DefaultInfo so that the macro can interop with `proto_library`)

NOTE: this is not a breaking change because the @bazel/labs package comes with no stability guarantee.
If you're using protobufjs_ts_library you can duplicate the macro from the example to keep using it.

refactor: use ProtoInfo in the protobufjs example

This ensures that users don't have to repeat the dependency tree