fix(builtin): always install source-map-support
We previously did this when patch_module_resolver was the default, because it had some extra (unrelated) logic at the end to install it.
When we flipped the default in 3.0, we accidentally dropped the feature.

This just reinstates it using the source-map-support/register feature.
Similarly to our prior implementation, we just call our own vendored source-map-support so users don't have to think about it.

Also fix missing docs for npm_package_bin attrs stdout,stderr,exit_code_out

Fixes #2520