feat(esbuild): add output_css flag to esbuild()
When esbuild encounters an 'import "./foo.css"' file in normal (non-splitting)
mode, it will place the contents of that file in a .css file next to the
main output .js file.

This patch adds a flag output_css, which when set to True, will declare
the created .css file as another output, so it can be consumed by other
rules.