fix: fetch node toolchain based on exec platform, not target platform
Currently rules_nodejs can't be used when cross-compiling to a different
platform, if that platform does not have a node toolchain available.
This means if you want to cross-compile a project that includes JS
files, you need to do it in multiple steps:

- build the JS files with the default host platform
- build the other products with a target platform specified, copying
the generated JS files in

The fact that it's currently set to target makes me wonder if this
use case was just not considered at the time, or whether there's some
other workflow that changing to exec will break. Thoughts?